requirements.txt
streamlit
pandas
scikit-surprise
joblib
numpy
